Refactored nodes, added some new.

+++ Source/MathEx/Private/MathEx.cpp
@@ -54,13 +54,13 @@ float FMathEx::RandomSameDigitsNumber(float Value)
 {
 	int32 DecimalValue = FMath::TruncToInt(Value);
 
-	int32 DecimalDigits = DecimalValue > 0 ? (int)FMath::LogX(10, DecimalValue) + 1 : 1;
+	int32 DecimalDigits = DecimalValue > 0 ? (int)FMath::LogX(10.f, (float)DecimalValue) + 1 : 1;
 	int32 FractionalDigits = FMath::IsNearlyZero(Value - DecimalValue) ? 0 : 2;
 
-	float Result = FMath::RandRange(0, (int32)FMath::Pow(10, DecimalDigits) - 1);
+	float Result = FMath::RandRange(0, (int32)FMath::Pow(10.f, (float)DecimalDigits) - 1);
 	if (FractionalDigits > 0)
 	{
-		int32 Divider = (int32)FMath::Pow(10, FractionalDigits);
+		int32 Divider = (int32)FMath::Pow(10.f, (float)FractionalDigits);
 		Result += (float)FMath::RandRange(0, Divider - 1) / Divider;
 	}
